Create 003_web_server_Python_v3.py
In this implementation, we use the BaseHTTPRequestHandler class provided by the http.server module as the base class for our custom request handler. By subclassing it and overriding the do_GET method, we define the logic to handle incoming GET requests. In this case, we simply send a basic HTML response with the message "Hello, World!".

We create an instance of the HTTPServer class, passing the server address and our custom request handler. The server is started with the serve_forever() method, which listens for incoming requests and handles them according to the logic defined in the request handler.

To run the web server, save the code in a file (e.g., web_server.py) and execute it using a Python interpreter. The server will listen on port 8000. Open a web browser and navigate to http://localhost:8000, and you should see the "Hello, World!" message displayed.

Please note that this implementation is still a basic web server and may not cover all possible scenarios or handle more complex features. For more advanced functionality, it is recommended to use more robust frameworks or libraries specifically designed for web development in Python, such as Flask or Django.